feat add File view
