+++ src/lib/File.svelte
@@ -1,0 +1,12 @@
+<script lang="ts">
+	import { page } from '$app/stores';
+
+	const { content } = $page.stuff;
+</script>
+
+<code>{content}</code>
+
+<style lang="sass">
+    code
+        white-space: pre
+</style>
